functions/requirements.txt
firebase_functions~=0.4.2
firebase-admin~=7.1.0
google-cloud-translate>=3.22.0
google-cloud-secret-manager>=2.25.0
google-cloud-storage>=3.1.1
openai>=1.50.0
requests>=2.31.0
